+++ weni/activities/signals.py
@@ -1,5 +1,3 @@
-import celery
-from django.db import models
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
@@ -8,21 +6,7 @@ from temba.flows.models import Flow
 from temba.triggers.models import Trigger
 from temba.campaigns.models import Campaign
 
-
-def create_recent_activity(instance: models.Model, created: bool):
-    if instance.is_active:
-        action = "CREATE" if created else "UPDATE"
-
-        celery.execute.send_task(
-            "create_recent_activity",
-            kwargs=dict(
-                action=action,
-                entity=instance.__class__.__name__.upper(),
-                entity_name=getattr(instance, "name", None),
-                user=instance.modified_by.email,
-                flow_organization=str(instance.org.uuid),
-            ),
-        )
+from weni.activities.recent_activities import create_recent_activity
 
 
 @receiver(post_save, sender=Channel)
@@ -31,7 +15,7 @@ def channel_recent_activity_signal(sender, instance: Channel, created: bool, **k
     if instance.channel_type not in ["WA", "WAC"] or update_fields != frozenset(
         {
             "config",
-        }
+        },
     ):
         create_recent_activity(instance, created)
 
@@ -61,5 +45,7 @@ def trigger_recent_activity_signal(sender, instance: Trigger, created: bool, **k
 
 
 @receiver(post_save, sender=Campaign)
-def campaign_recent_activity_signal(sender, instance: Campaign, created: bool, **kwargs):
+def campaign_recent_activity_signal(
+    sender, instance: Campaign, created: bool, **kwargs
+):
     create_recent_activity(instance, created)