+++ weni/internal/msgs/tasks.py
@@ -1,5 +1,8 @@
+from datetime import datetime, timedelta
 import smtplib
 import logging
+from collections import defaultdict
+from dateutil import parser
 
 from celery import shared_task
 
@@ -24,46 +27,76 @@ logger = logging.getLogger(__name__)
 @shared_task(name="generate_sent_report_messages")
 def generate_sent_report_messages(**kwargs):
     org_id = kwargs.get("org_id")
-
     data = kwargs.get("data")
-    start_date = data["start_date"]
-    end_date = data["end_date"]
+    start_date = parse_date(data["start_date"])
+    end_date = parse_date(data["end_date"])
 
-    query = f"""
-        SELECT
-            template.name AS "Template",
-            flow.name AS "Fluxos Utilizados",
-            COUNT(msg.id) AS "Total por Template"
-        FROM
-            public.msgs_msg AS msg
-            INNER JOIN public.templates_template AS template
-                ON CAST(template.uuid AS text) = msg.metadata::json -> 'templating' -> 'template' ->> 'uuid'
-            INNER JOIN public.flows_flow_template_dependencies AS depent
-                ON depent.template_id = template.id
-            INNER JOIN public.flows_flow AS flow
-                ON flow.id = depent.flow_id
-        WHERE
-            msg.created_on BETWEEN '{start_date}' AND '{end_date}'
-            AND msg.metadata::jsonb -> 'templating' IS NOT NULL
-            AND msg.org_id = {org_id}
-            AND msg.status IN ('S', 'D', 'V')
-        GROUP BY
-            template.name, flow.name
-        ORDER BY
-            COUNT(msg.id) DESC;
-    """
-    filename = "Mensagens Enviadas.xlsx"
+    logger.info(f"Starting report messages to org {org_id}")
+
+    lock_key = f"template-messages-lock:{org_id}"
+    lock_ttl = 10800  # TTL of 3 hours in seconds
 
     try:
         redis_client = get_redis_connection()
-        query_data = fetch_query_results(query)
-        processed_query_data = process_query_results(query_data)
+        redis_client.set(lock_key, "locked")
+        redis_client.expire(lock_key, lock_ttl)
+
+        logger.info(f"Starting task: generate_sent_report_messages for org {org_id}")
+
+        time_batches = partition_date_range(start_date, end_date, timedelta(hours=4))
+        batch_data = []
+
+        for batch_start, batch_end in time_batches:
+            query = f"""
+                SELECT
+                    template.name AS "Template",
+                    flow.name AS "Fluxos Utilizados",
+                    COUNT(msg.id) AS "Total por Template"
+                FROM
+                    public.msgs_msg AS msg
+                    INNER JOIN public.templates_template AS template
+                        ON template.name = msg.template
+                    INNER JOIN public.flows_flow_template_dependencies AS depent
+                        ON depent.template_id = template.id
+                    INNER JOIN public.flows_flow AS flow
+                        ON flow.id = depent.flow_id
+                WHERE
+                    msg.created_on BETWEEN '{batch_start}' AND '{batch_end}'
+                    AND msg.template IS NOT NULL
+                    AND msg.org_id = {org_id}
+                    AND msg.status IN ('S', 'D', 'V')
+                GROUP BY
+                    template.name, flow.name
+                ORDER BY
+                    COUNT(msg.id) DESC;
+            """
+            query_data = fetch_query_results(query)
+            batch_data.extend(query_data)
+
+        processed_query_data = process_query_results(batch_data)
         file = export_data_to_excel(processed_query_data)
-        send_report_file(file_stream=file, file_name=filename, data=data)
+        send_report_file(file_stream=file, file_name="Mensagens Enviadas.xlsx", data=data)
+        logger.info(f"Message Reports sent to org {org_id}")
+
     except Exception as e:
-        logger.info(f"Fail to generate report: ORG {org_id}: {e}")
+        logger.error(f"Fail to generate report for org {org_id}: {e}")
     finally:
-        redis_client.delete(f"template-messages-lock:{org_id}")
+        redis_client.delete(lock_key)
+        logger.info(f"Lock released for org {org_id}")
+
+
+def parse_date(date_str):
+    dt = parser.parse(date_str)
+    return datetime(dt.year, dt.month, dt.day)
+
+
+def partition_date_range(start_date, end_date, delta):
+    adjusted_end_date = end_date + timedelta(days=1)
+    current_date = start_date
+    while current_date < adjusted_end_date:
+        batch_end = min(current_date + delta, adjusted_end_date)
+        yield current_date, batch_end
+        current_date = batch_end
 
 
 def fetch_query_results(query):
@@ -74,20 +107,15 @@ def fetch_query_results(query):
 
 
 def process_query_results(data):
-    processed_data = []
-    template_dict = {}
+    grouped_data = defaultdict(int)
 
     for row in data:
         template_name, flow_name, total = row
+        key = (template_name, flow_name)
+        grouped_data[key] += total
 
-        if template_name not in template_dict:
-            template_dict[template_name] = flow_name
-            processed_data.append((template_name, flow_name, total))
-        else:
-            processed_data.append((template_name, flow_name))
-
+    processed_data = [(template, flow, total) for (template, flow), total in grouped_data.items()]
     return processed_data
-
 
 def export_data_to_excel(data):
     workbook = Workbook()